--- src/main/java/Priority.java
@@ -1,5 +1,0 @@
-public enum Priority {
-    Minor,
-    Critical,
-    Normal,
-}
