--- src/main/java/Store.java
@@ -1,6 +1,0 @@
-public interface Store {
-    public void addItem(String s);
-    public void changeStatus(int number, Status st);
-    public void removeItem(int number);
-    public void printAll();
-}
