--- src/main/java/Tag.java
@@ -1,6 +1,0 @@
-public enum Tag {
-    DailyRoutine,
-    Home,
-    Work,
-    Reading,
-}
