+++ src/main/java/com/todolist/app/model/Command.java
@@ -1,3 +1,5 @@
+package com.todolist.app.model;
+
 /**
  * These are commands, which are available for using
  */
