+++ src/main/java/com/todolist/app/model/Item.java
@@ -1,3 +1,6 @@
+package com.todolist.app.model;
+import com.todolist.app.model.Tag;
+
 import java.util.ArrayList;
 import java.util.List;
