+++ src/main/java/com/todolist/app/model/Priority.java
@@ -1,0 +1,7 @@
+package com.todolist.app.model;
+
+public enum Priority {
+    MINOR,
+    CRITICAL,
+    NORMAL
+}
