+++ src/main/java/com/todolist/app/model/Status.java
@@ -1,3 +1,5 @@
+package com.todolist.app.model;
+
 /**
  * These are statuses, which user can mark his tasks
  */
@@ -5,17 +7,17 @@ public enum Status {
     /**
      * This status means that the task is done
      */
-    Completed,
+    COMPLETED,
     /**
      * This status means that the task hasn't been done yet
      */
-    Incompleted,
+    INCOMPLETED,
     /**
      * This status means that the task is being worked
      */
-    InProgress,
+    INPROGRESS,
     /**
      * This status means that the task is in a queue
      */
-    Pending,
+    PENDING
 }
