+++ src/main/java/com/todolist/app/model/Tag.java
@@ -1,0 +1,8 @@
+package com.todolist.app.model;
+
+public enum Tag {
+    DAILYROUTINE,
+    HOME,
+    WORK,
+    READING
+}
