+++ src/main/java/com/todolist/app/model/ToDoCommand.java
@@ -1,3 +1,5 @@
+package com.todolist.app.model;
+
 public class ToDoCommand {
     private int number;
     private Command command;
