+++ src/main/java/com/todolist/app/servlet/ToDoServlet.java
@@ -1,3 +1,5 @@
+package com.todolist.app.servlet;
+
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
 import jakarta.servlet.annotation.*;
