+++ src/main/webapp/task-style.css
@@ -20,7 +20,7 @@ body{
     left: 50%;
     margin-right: -50%;
     transform: translate(-50%, -50%);
-    width: 1085px;
+    width: 490px;
     height: 540px;
     padding: 50px 70px;
     border-radius: 25px;
@@ -53,7 +53,7 @@ body{
     font-family: sans-serif;
     border-radius: 10px;
     margin-top: 5px;
-    margin-left: 195px;
+    margin-left: 192px;
     width: 75px;
     height: 30px;
 }
@@ -79,6 +79,14 @@ body{
     padding: 10px 0;
 }
 
+.newtask .form-group-status label{
+    font-size: 12px;
+    font-weight: bold;
+    font-family: sans-serif;
+    margin-top: 15px;
+    padding: 10px 0;
+}
+
 .newtask .label-form{
     font-size: 13px;
     font-family: sans-serif;
@@ -93,107 +101,52 @@ body{
     margin-bottom: 5px;
 }
 
-.newtask .checkbox{
-    border-radius: 10px;
-    box-shadow: 0 0 1em lightgrey;
+
+
+.newtask .select {
+    color: black;
     background-color: white;
-    right: 0px;
-    width: 200px;
-    height: 150px;
-    padding: 0;
-}
-
-.newtask .checkbox-status{
-    border-radius: 10px;
-    box-shadow: 0 0 1em lightgrey;
-    background-color: white;
-    right: 0px;
-    width: 200px;
-    height: 150px;
-}
-
-.newtask .checkbox-status label{
-    padding: 15px 0;
-    padding-left: 15px;
-    margin-top: 15px;
-}
-
-.newtask .tags{
-    border-radius: 10px;
-    box-shadow: 0 0 1em lightgrey;
-    background-color: white;
-    right: 0px;
-    width: 200px;
-    height: 150px;
-    padding: 0;
-}
-
-.newtask .checkbox-priority{
-    border-radius: 10px;
-    box-shadow: 0 0 1em lightgrey;
-    background-color: white;
-    right: 0px;
-    width: 200px;
-    height: 115px;
-    padding: 10px 0;
-}
-
-.newtask .label-checkbox{
+    cursor: pointer;
     font-size: 13px;
     font-family: sans-serif;
-}
 
-.newtask .tags label{
-    float: bottom;
-    padding-left: 15px;
-    padding-top: -15px;
-}
-
-.newtask .checkbox-priority label{
-    padding: 15px;
-}
-
-.newtask .div-label-status{
-    padding-top: 10px;
-    margin-bottom: 10px;
-}
-
-.newtask .label-tags{
-    padding-bottom: 10px;
-    margin-bottom: 10px;
-}
-
-.newtask .div-label{
+    border-radius: 5px;
     margin-bottom: 20px;
+    width: 350px;
+    height: 30px;
+    border-color: black;
+    border-width: 2px;
+    border-style: solid;
+    margin-top: 5px;
 }
 
-
-
-input[type=radio]{
-    vertical-align: middle;
-    float: right;
-    margin-right: 15px;
-}
-
-input[type=checkbox]{
-    vertical-align: middle;
-}
-
-
-.newtask .label-form-priority{
+.newtask .select-mult {
+    color: black;
+    background-color: white;
+    cursor: pointer;
     font-size: 13px;
     font-family: sans-serif;
-    padding: 5px 0;
-    right: 0px;
-    margin-left: 450px;
-    margin-top: 15px;
+
+    border-radius: 5px;
+    margin-bottom: 20px;
+    width: 350px;
+    height: 30px;
+    border-color: black;
+    border-width: 2px;
+    border-style: solid;
+    margin-top: 5px;
 }
 
-.newtask .label-form-tags{
-    font-size: 13px;
-    font-family: sans-serif;
-    padding: 5px 0;
-    right: 0px;
-    margin-left: 730px;
-    margin-top: -178px;
-}
+
+
+
+
+
+
+
+
+
+
+
+
+
